MB-55027: Buffer mcbp stats responses from background tasks
Previously, instances of StatsTask would buffer up all of the stats in a
vector<pair<string, string>> and write the stats one by one to the
connection after the task is done.

Changed the StatsTask to buffer up the actual MCBP responses in a
single IOBuf, instead of a vector.

Also changed the StatsTask interface to use this new callback and to write
the stats itself from drainBufferedStatsToOutput instead of deferring to
the caller.

The plan is to change StatsTask to drainBufferStatsToOutput regularly,
every time the buffer is filled, and to yield to the front-end until
the IOBuf is written to the socket.

Change-Id: I9ee491169e1a055e769cbbc65fdc4786133b13eb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199277
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>