MB-55027: Do not generate stats for dead streams
DCP supports the control message send_stream_end_on_client_close_stream
which requires us to acknowledge the stream ending when the client
requests it.

However, we've seen cases in which due to networking or bugs in another
service, a consumer can send CloseStreams for a large number of streams,
when the connection buffer is already full, and do that in a loop.

In that case, we can keep opening new streams, but cannot free up the
old streams because we cannot send the stream end message. Then, with
a large number of streams, we can generate very large cbstats payload
that can timeout, so we get no observability on the issue.

While we're not fixing the cause of this issue (which is network/other
service), avoid a situation in which we cannot get our stats, by not
generating stats for dead streams.

Change-Id: I29ff6b6eadf23c2e55e7f13f8205460cb3a08723
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199363
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>