MB-59286: Warmup: replace AtomicHashMap with Synchronized<>
Warmup::warmedUpVbuckets currently uses a folly::AtomicHashMap for
tracking the vBuckets which are being warmed up. This class does allow
lock-free concurrent access to the map, however it comes with
additional complexity including singleton memory allocations in the
implementation details of AtomicHashMap. Those singlestons are
problematic for memory tracking, as they will be accounted to the
bucket which is running when the map is first accessed. While we could
explicitly pre-create the AtomicHashMap outside a bucket context, that
adds additional complexity; and in this case we don't really need the
performance of AtomicHashArray - the map is essentially populated once
(by an IO-bound task), and then elements infrequently accessed during
the rest of warmup.

As such, simplify to use a std::unordered_map wrapped in Synchronized
for thread-safe access (from the multiple warmup tasks). This avoids
the whole memory accounting issue.

Change-Id: I4997bef2609aecb95341e4ccf70514893af48153
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199651
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>