MB-52682: Fix item_eviction_initial_mfu_percentile default
During code review, noticed that the initial MFU percentile for the
upfront_mfu_only eviction strategy is set to 50. This is an oversight
between patchset 31 and 32, where I incorrectly pushed 50 as the
default. Also, a branch checking for overflow was also omitted from the
final change.

*Background*

When originally implementing this in
https://review.couchbase.org/c/kv_engine/+/181156, I tried multiple
different values, including 50, however, ultimately settled on 25, which
is also referenced in the HiFi_MFU_eviction_policy.md and in comment on
the ticket.

*Reasoning*

A single run of the ItemPager might need to evict anywhere between
0-100% of the items in memory (depending on other memory consumers and
eviction policy). For the case where 50-100% of the bucket quota is
evictable, creating new items at the 25%ile means the item pager will
be able to evict the new items either on the first pass or the second one.

This is because at each pass the item pager will evict ~12% (ratio
between high_wat and low_wat) of the current set of items in the
HashTable, by picking the 12%ile from the MFU histogram.

No difference was observed in the tests that we have, between 25%ile and
50%ile however.

Change-Id: I0c469a006c439b471974afc3ba385d567ebbe60b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199450
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>