MB-59287: Restore missing periodic wakeup for ItemPager
The ItemPager task can be woken in one of two ways:

1. If a memory condition is detected (mem_used > high_watermark).
2. Periodically, by default every 5s.

Method (1) _should_ ideally be sufficient to trigger the ItemPager,
but we only check on specific paths (e.g. checkAndMaybeFreeMemory),
which currently does not include lookup-only methods. As such, for
100% read workloads which are <100% resident (and hence will need to
BGFetch items into memory and hence increase mem_used), we rely on
method (2) to trigger memory recovery via ItemPager.

However, method (2) was inadvertently disabled was part of "MB-59062
[1/7]: Refactor ItemPager in preparation for EpTask" (cd3c048bb) - the
override of getSleepTime() was incorrectly removed from ItemPager
class, as it was (mistakenly) assumed that the initialSleepTime was
reused for subsequent sleep time. This results in the parent
NotifiableTask::getSleepTime() method being used instead, which
returns INT_MAX (forever) as the sleep time.

As a consequence we end up exeeding the high watermark (and the 93%
tempOOM watermark) for 100% read workloads, essentially livelocking.

Fix by restoring the getSleepTime override and add regression test.

Change-Id: Iefe00d25cf15b360f1e09ab7471b0b5f95be26ea
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199574
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>