MB-58868: Make AS and Filter addStats return unprefixed stats
And do the stream name prefixing in Producer::addStats.

We don't want individual per-stream keys prefixed with the stream name.
In a follow up patch, I will make all the per-stream stats be collected
in a JSON format and exposed as a single stat.

The AS filter stats are prefixed with 'filter_' now and returned from
AS::addStats.

Change-Id: I402bf96258e33e31c608469f4a97ae4fc542b35b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199630
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>