MB-58644: ReplicaEvictedBeforeActive test: Make robust to fluctuations in mem_used
This test attempts to populate a number of replica and active items
such that, when item paging runs, the replica items _should_ be
sufficient to recover all desired memory - i.e. difference between
current mem_used and low watermark should be less than the memory we
can recover from just replica items, meaning no actives need to be
evicted.

However, the test is somewhat brittle as it attempts to exactly set
the low watermark to the difference between 'current' mem_used replica
pageable memory. The problem with this is that 'current' mem_used can
chagne between this calculaton and when the PagerTask actually
runs. This is sometimes masked by the fact we request estimated
mem_used values, but when using more strict debug memory tracking
(e.g. JEArenaSimpleTracker), we can see enough of an increase in
mem_used between setting low watermark and paging running, that the
low watermark is "too low" and requires some active items to also be
paged out.

Address this by setting the low watermark a little higher - such that
only 75% of the replica items need to be evicted to recover sufficient
memory. This means that even if mem_used does increase by a small
percentage when the ItemPager runs, we have a margin of error
(i.e. perhaps we evict 76, 78% of replica items), but still no active
items.

Change-Id: Ic6f7b646547cd4593e8ab1b76b7bd3ae51e22c39
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199669
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>