MB-59368: Use getPageableMemCurrent/MemLowWatermark in PagingVisitor
The MB highlighted that for ephemeral auto_delete, paging can be
triggered as the trigger uses getPageableMemCurrent/MemLowWatermark
but the PagingVisitor immediately cancels paging because it compared
getEstimatedTotalMemoryUsed vs stats.mem_low_wat, nothing is freed
and paging is immediately triggered to run again - resulting in a
noticeable increase in CPU but no memory reduction.

Commit changes PagingVisitor to use the Pageable functions when
deciding to stop, memory is now released and no further paging is
triggered.

Change-Id: I2668c55df443ff907659bda20f38c15af6df9dbe
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199899
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>