MB-58961: Update AS::lastReadSeqno for empty in-memory snapshot
Let's consider this case:
- 2 collections, default + fruit
- a single ActiveStream that filters on fruit

In the case where the stream ends up processing a checkpoint like:

  [e:4 cs:4 m(keyD4):4 m(keyD5):5)

(ie, all mutations in the checkpoint are in collection default and so
they are filtered out at stream), then we do move the stream cursor in
checkpoint but we miss to update lastReadSeqno=5.

That misbehaviour creates the preconditions for a possible subsequent
error that might end up in a Producer disconnection.
That is detailed in the new test in this patch.

Change-Id: Ia0b097f19286c5e1f50a82eb0fee841e722f89b0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199966
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>