Add kv_stats_log.py for easier grepping of stats.log
A python script which parses stats.log and can emit greppable output.
This is achieved by prepending each line of the output by the cbstats
command and bucket name (-g option).

  cbstats all:default: auth_cmds:   	0
  cbstats all:default: auth_errors:	0

By passing in -c/-a/-b a specific command output can be extracted. For
example, using:

  kv_stats_log.py stats.log -c cbstats -a kvtimings -b default

We get the output of cbstats kvtimings for the bucket default only.

The script achieves reasonable throughput, for a 62 MiB stats.log file
on my machine, with -g, it outputs:

  kv_stats_log.py took: 0.283s, Read: 61.00 MiB, Rate: 215.46 MiB/s

There is already a lot of stats.log parsing functionality in nutshell
(mcdstatsparser.py), however that parses specific stats from the file
and creates an in-memory representation.

This script performs very lightweight processing, only reading a chunk
from the stats.log at a time and prints it as requested.

Change-Id: Ib557abee1a85b7c93142345627b30168ad93bf71
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/200044
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>