Strip off (and ignore) extra NUL provided in plain auth message
According to the RFC the auth message looks like:

   message   = [authzid] UTF8NUL authcid UTF8NUL passwd
   authcid   = 1*SAFE ; MUST accept up to 255 octets
   authzid   = 1*SAFE ; MUST accept up to 255 octets
   passwd    = 1*SAFE ; MUST accept up to 255 octets
   UTF8NUL   = %x00 ; UTF-8 encoded NUL character

Some clients did however send an additional NUL after
the password. This fix a "regression" introduced in
the simplification of parsing of the PLAIN message.

I'm writing "regression" as it used to allow an
incorrect encoding in the packets and given that
we don't know which clients are vulnerable to this
we should probably continue to allow this invalid
encoding

Change-Id: I7ef57f8e7649d262ad4bbd245a7def2993744866
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/200115
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>