MB-58088: Introduce the NOT_LOCKED (0x0e) MCBP status code
…and return not_locked when unlocking a not locked document.

The status code indicates a possible application logic error, and not a
transient server state.

Previously we were returning temporary_failure, so not_locked is mapped
to temporary_failure for clients not supporting XERROR.

Clients with XERROR support will receive the new status code, and will
use the JSON error map to interpret it. As we set only the "item-only"
attribute, no special handling is expected. In particular, clients are
not expected to retry the operation or return a timeout failure.

During testing, I found that SDKs interpret the attributes
inconsistently, and it was not possible to emulate the existing
exception behaviour in all the SDKs. Given that this would have been
a breaking change anyway, we opted to set the semantically correct
attributes.

Change-Id: I30a717dde233ae4188d46c0ede3a20d35dce7baa
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198641
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>