MB-55027: EWB until the stats output is written to the socket
If the command yields, we should write the buffered stats to the socket
and become EWB.

The stats are wrapped in a SendBuffer, which will notifyIoComplete the
cookie when the underlying libevent/folly connection has written it.

We will continue execution when the cookie is notified.

The next patch will allow the engine to check when we need to yield,
which will be when the buffered stats reach a threshold.

Change-Id: I7dda415faba093713ddf88b79ac68b16e0babf2a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/200376
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>