MB-59518: Stop PassiveStream from processing commit against active VB
Add VBucket state checks to processCommit so that the commit is not
processed against an active vbucket.

Without the state check a buffered commit could be processed
concurrently with a replica->active state change. A unit test is added
which reproduced the issue by buffering a commit and using a test hook
to change state whilst the buffered data is processed.

Note this bug extends to:

* abort
* mutation, deletion, system-event
* snapshot marker

Subsequent patches address these operations.

Change-Id: Ifb5e1c7703d7fe2af0b92b699dffb091dea0b235
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/200568
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker