MB-59310: Make ActiveStream resilient to empty checkpoints
In MB-59282 and MB-57443 we hit two cases of external DCP components
wrongly streaming from replica vbuckets.
That revealed an issue with the ActiveStream code, where the stream
might end up running the ActiveStreamCheckpointProcessorTask in a state
where the CM doesn't have any items for cursor. The stream code isn't
relisient to that, so an exception is triggered and the connection
closed.

It is unclear what causes the extra Task execution. The Task's wake up
logic should trigger the Task execution only where there's at least 1
item for the DCP cursor (any item, 1 single meta-items is enough).

That CM state is benign, so this patch simulates that scenario in a
unit test and makes ActiveStream resilient to that case.
Rather than throwing, the stream now logs a warning and continues the
execution. That prevents disconnections.

Change-Id: Iefecacf253e88f25e702da5b2616c6a368f6620e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/200729
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>