MB-59189: Do not access the cursor without a queueLock
Logging added under MB-58961 in neo accesses the cursor's position
without a checkpoint manager lock, which is unsafe, as the checkpoint
manager might have to dedup the item this cursor is currently on at any
time.

Return the Item the cursor gets registered on in CursorRegResult, so the
calling code can access it without going via a cursor deref to print the
info log.

Change-Id: Ie819af7fd12743d18579f17652eb00cf47e55c9e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/200830
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>