MB-56646: Reliably notify cookies waiting for warmup during shutdown
Some operations can end up blocked if warmup is incomplete, this is to
ensure that all vbuckets are loaded before such operations execute.
For example setVbucketState.

MB-56646 identfied that GetFailoverLog when executed from a DCP
connection could later crash the server if shutdown is requested
whilst blocked on warmup. This is because the DCP connection was
disconnected by the initiate_shutdown function yet the cookies
(raw pointer) which Warmup is holding onto are not notified until
the next part of shutdown cancel_all_operations_in_ewb_state. This
could leave Warmup with a dangling Cookie* that it will call
notifyIOComplete on.

Another subtle issue is that the existing code in
cancel_all_operations_in_ewb_state which notifies the cookies is
unreliable, it would only trigger whilst Warmup is !finishedLoadingData
yet that flag can be set to true whilst Warmup is still running again
allowing for a dangling pointer.

Change-Id: Ie2ca8f91095188545bc4941582bb669e7582eece
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/200941
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>