MB-58350: Wait for clients to disconnect before deleting bucket
As part of the fix for MB-56067, we converted the blocking bucket
deletion code into a state machine.

The state machine should only make progress if the bucket has no clients
(connections) associated with it.

However, there appears to be a bug in the state machine code, where we
incorrectly move the state machine to complete bucket deletion, before
all clients have disconnected.

This will happen 2 minutes after the DELETE_BUCKET started executing,
due to a missing check for the current number of clients in that code
path.

As a consequence of this, we crash in Connection::propagateDisconnect().

Change-Id: I1aaffb2481107221020974ba7729224e4b764f4b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/201110
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>