MB-59710: DCP connections must wait for blocked cookies
DCP connections didn't correctly use the 'ewouldblock' state
variable in the initial refactors from the old state machine
over to using bufferevents. Over the time we've cleaned up
the usage of the variable and introduced an assertion as
part of shutting down the cookie that it is no longer
in a blocked state.

We've had a lurking bomb in our code for a while as we would
be using dangling pointers for DCP connections during shutdown
if the core was executing an incomming command at the same time
as we where trying shut down the connection.

This patch waits for all blocked cookies to complete
before continuing the shutdown (DCP connections use the
reference count and scheduleDcpStep and not the blocked
state of the cookie)

Change-Id: I27c401c67a7b1d54d542f39df768842ddfdac831
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/201211
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker