MB-59452: Don't use xattrs for collection related metadata
The addition of modify collection in 7.2 required some extra work to
keep support for "legacy" DCP users, that is to not expose the seqno
of the modify event through various APIs (e.g. getAllSeqnos).

The orginal change did this by stashing this one-off seqno into an
xattr blob attached to the value of the modify event. Thus only
the modify of the default collection ever stored this extra data.

MB-59452 reminded us that xattrs can introduce a performance cost and
the modify of the default collection should not be the trigger for
a user to pay that cost.

Fix by moving the extra metadata into the flatbuffer data so no xattr
values are generated by KV itself.

1) The modify collection "write" path now changes to always stash the
special seqno inside the flatbuffers.

2) The warmup path checks for xattr and can load both encodings.

Note there is no explicit "sanitise" path. DCP never transmitted the
xattr part of the modify and the special seqno was always recalculated
by the recipient. In this case a rebalance will result in the xattr
blob being removed and never stored again.

Change-Id: I33872e304f0ec53ec2ba3de8cf161546ba12ff0c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/201296
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>