MB-56917: Refactor getMulti to accept a createItemCallback
Passing a callback allows access to engine specific attributes/methods,
e.g. MemoryTracker and createItem, within API calls to disk. The
getDefaultCreateItemCallback returns an implementation of the original
behaviour of the system, and createItemCallback implements a check for
sufficient memory before creating a new item.

NOTE: This patch only refactors getMulti to accept a callback but it is
not used.

Change-Id: I4280947283fd37de7e3ba2833a134649f735c89a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/201234
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>