MB-59186 Use new Status::Code::OkDocNotFound
Magma's GetLocal, GetBySeqno, Get and GetDocs use a combination of
Status and a bool to indicate result of request processing. The Status
indicates whether the request processing itself succeeded or not. If it
did, then the bool parameter indicates whether the doc was found or not.

For cleaner code, we now refactor the code to only depend on Status to
indicate all situations. For this we introduce a new
Status::Code::OkDocNotFound which is a subcode of Status::Code::Ok and
indicates that request processing succeeded but no doc was found.

Linked magma change:
https://review.couchbase.org/c/magma/+/200978

Change-Id: I0576993d8e92345045afb418bc9ae5683d3a5637
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/201289
Tested-by: Rohan Suri <rohan.suri@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>