MB-59474: Fix race causing DCP BufferAcks for 0 bytes
The code in FC::handleFlowCtl checks whether the number of ackable bytes
is non-zero and time has elapsed OR whether that number is above the
ack threshold.

If either condition is true, it updates the internal counters and sends
an acknowledgement with the original value that was read.

1. Check non-zero and time has elapsed
2. Check above ack threshold

The code used to do an atomic read of the same variable for each of the
checks, which meant that checks potentially see different values. If
then used the first value that was read to update the FlowControl
buffer.

This creates a race where if the number of ackable bytes changed between
1. and 2., we could end up sending a BufferAck for 0 bytes (and updating
the flow control buffer by 0/noop). This does not result in incorrect
internal state.

The mcbp protocol does not require that BufferAcks are non-zero, but it
causes some tests to fail.

Change-Id: If29fa57999f92ba5b6e2c8e1b3ddf812003b8384
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/201328
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>