Log the bucket state for SetClusterConfig
We can't set the bucket cluster configuration unless the bucket
is in a "ready" state. In this case we would return "temporary
failure" back to the caller (ns_server). Previously it wasn't
clear what the state of the bucket was in these cases, so
this patch logs the state of the bucket in addition to the
rest of the error information.

Change-Id: Ia64a554a38ab624e0fca3a1eef94859c267406a8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/201614
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>