MB-59883: Remove support for 'recycle' connection mode
The server can't (currently) distinguish a working connection from a broken connection. It just has a list of the least recently used connection. Clients typically creates a pool of connection used to send data which means that the connection pool logic would be broken by this design. Worst case scenario in this mode is that all connections are constantly recycled and no-one is able to perform any operations.

Change-Id: I19b8b30e50349a25e99db1000f8cb7c06904935e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/201794
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>