MB-58868: Add alternative format for STAT(dcp) and use for cbstats
The 'dcp' statistics group now supports an additional attribute used to
specify the format of producer/consumer stream statistics.

With this patch, in addition to the older format, one can now request
stream stats to either be ignored or formatted as JSON, which is more
efficient.

Results from local testing with dcpdrain (1024 streams) and requesting
the dcp stats group using cbstats with JSON/legacy format (real time,
95% confidence, 5 samples):

 - legacy	0.2186s ±0.039 (x1)	data:1746 KiB
 - json		0.1242s ±0.033 (x0.56)	data:662 KiB (62% smaller)

That is, the command `cbstats dcp >/dev/null` executed almost twice as
fast, with 1024 ActiveStreams created by dcpdrain.

We see this scales much better with the number of streams, partly
because of inefficiency in cbstats (output stored in Python dict).

Test with 50 dcpdrain instances (50K streams):

 - legacy	5.414s (x1)	data:85 MiB
 - json		0.675s (0.12x)	data:32 MiB (62% smaller)

Additionally, tested with 50K streams and mcstat, which currently
doesn't support passing the additional stream_format argument and uses
the non-JSON format.

 - mcstat	1.432s		data:66.6 MiB

So mcstat output is smaller than the legacy output from cbstats, because
of less formatting, but larger than the JSON format (and slower than
cbstats in that case).

Change-Id: I449ca438402f815e0ae2feacf3e79c37128a51c6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/202369
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>