Fix inefficient memory allocation when stats exceed 16k
The code use a single iobuf and call reserve() for each
new stat being added causing a rellocation when the data
won't fit in the buffer.

A lot of the stat groups emit more than 16k of data causing
a lot of reallocations and memory movement. Instead of
reallocating the buffer we can just allocate a new buffer
and send all of them when we're done

Change-Id: I39e8a757b83c7359e3ff66265760e0468ca2a1ec
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/202427
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>