MB-60046: Ephemeral: Don't auto-delete committed and pending keys.
Prior to this change the auto-delete path was able to generate a delete
of the committed item whilst there is a pending write. The delete would
be rejected by the consumer and disconnect DCP (causing a cycle of
disconnect).

This change updates EphemeralVBucket::eligibleToPageOut to check the
current hash-table bucket for a pending key.

This approach is deliberately simple and the inefficient extra search of
the hash-chain is tolerated to provide a rapid solution to this issue.

Change-Id: If83ecb927026a4ef4ca2dff813a8729ec67da211
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/202456
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>