MB-56917: Refactor fetchDoc to accept a createItemCB
Allow fetchDoc to accept the callback from getMultiCallback such that
it can be passed into makeItemFromDocInfo and called when an Item is
created.

This patch also creates a default callback
(getDefaultCreateItemCallback) that implements the original behaviour,
i.e. without any memory checks, and used when fetchDoc is called within
getWithHeader, getBySeqno and unit tests that do not have an engine.

Change-Id: I01581d222065bd6b132b19d1e20ec725a29e5ef6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/201676
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>