MB-56917: Check memory before creating a new item from BGFetch
makeItemFromDocInfo accepts a callback that is now used to create
a new Item in Couch-KVStore. For example,
getMultiCallback->fetchDoc passes a callback that will only
create a new item if it will not exceed the mutation watermark.
This means that, items in the BGFetch queue are processed
sequencially, if it will exceed the mutation watermark
temp_failure is returned, otherwise it will succeed.

Change-Id: I5871a26c228b27a4504fcc9db29684d280cc384a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/202554
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>