MB-56917: Check memory before creating a new item from BGFetch
makeItemFromDocInfo accepts a callback that is now used to create
a new Item in Couch-KVStore. For example,
getMultiCallback->fetchDoc passes a callback that will only
create a new item if it will not exceed the mutation watermark.
This means that, items in the BGFetch queue are processed
sequencially, if it will exceed the mutation watermark
temp_failure is returned, otherwise it will succeed.

Change-Id: I0816b782d677520af658d2bcecf8d0a6a5431ea0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/203123
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>