MB-59855: Do not DCPBackfill::scan() on create for Sequential backfills
In MB-52975, we folded the Create and Scan phases of backfills, such
that a backfills always runs a Scan phase after it opens a disk
snapshot.

This breaks the expectation of schedule_order=sequential. To work around
this issue, while preserving the reduced backfill time achieved by
MB-52975, add an exemption for sequential backfills (so round-robin
backfills still scan() immediately after a create()).

Added DCPBackfillCreateMode which is set to CreateOnly for sequential
backfills and CreateAndScan for round-robin. This parameter controls
whether a backfills should run a Scan phase after it opens a disk
snapshot.

Change-Id: I27b6d034f83515c2e87b9d55e580609989acd808
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/202791
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>