MB-60276: [BP] Avoid unnecessarily calling CouchKVStore::getDbFileInfo()
Backport to 7.2.5.

This method requires reading metadata from disk on Couchstore, which is
naturally very slow. We read this metadata for every STAT("all")
request, which is wasteful and can result in very long runtimes.

This patch avoids the call made in getHistoryDiskSize() under
Couchstore, as it doesn't support history in the first place.

Magma stores the file metadata in memory, so it is not affected.

Change-Id: I59f70b6227d511ff5fa055dc9d6d6cb9c4545a07
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/205316
Well-Formed: Restriction Checker
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Dan Owen <owend@couchbase.com>