MB-60276: [BP] Avoid unnecessarily calling CouchKVStore::getDbFileInfo()
Backport of 39798953 for 7.2.4 branch

This method requires reading metadata from disk on Couchstore, which is
naturally very slow. We read this metadata for every STAT("all")
request, which is wasteful and can result in very long runtimes.

This patch avoids the call made in getHistoryDiskSize() under
Couchstore, as it doesn't support history in the first place.

Magma stores the file metadata in memory, so it is not affected.

Change-Id: I60a83f744fa278ef84667af6123a3c1431453d6f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/208724
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Well-Formed: Restriction Checker