MB-56917: Check memory when fetching item from disk in Magma
Allow MagmaKVStore::makeItem to use createItemCb to determine if
sufficient memory is available to create a new item. If there is
available memory, then a GetValue with a pointer to a new item
and an engine_errc::success is returned. Otherwise, a GetValue
with nullptr is returned as an item and
engine_errc::temp_faillure / no_memory.

Change-Id: I0dd236fb7889d7ccc8858a215c102377f38a5005
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/203384
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>