MB-60858: [xattr] Add support for binary values
When the client set the flag SUBDOC_FLAG_BINARY_VALUE the
server encodes the value into the JSON as a string value
with the following encoding:

    cb-content-base64-encoded:base64encodedvalue

If the client tries to use SubdocReplaceBodyWithXattr with
the value SUBDOC_FLAG_BINARY_VALUE the value gets replaced
with the decoded value.

Clients who operate on binary values must also provide
SUBDOC_FLAG_BINARY_VALUE when requesting a binary value
to tell the server to convert the value back to a binary
value. If the value isn't a binary value the server will
fail the GET with SubdocFieldNotBinaryValue

If the client request the object without the flag the
encoded version of the data will be returned (same
would happen if the returned value is an object which
again contains a binary value)

   upsert 'tnx.op.staged binary-value, tnx.op.client foo'
   lookup "tnx"

    {
      "op": {
        "client": "foo",
        "staged": "cb-content-base64-encoded:YmFzZTY0X2VuY29kZV92YWx1ZQ=="
      }
    }

Change-Id: I8e48bd43b80a902c19a3460ee5d0435873d969ee
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/206239
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>