MB-59759: Do not send snap_start_seqno lower than the requested
In ActiveStream::snapshot(), where we create a snapshot marker before we
push items reads from the CM to the stream's readyQ, we have some logic
to pin the snapshot start seqno to the checkpoint.

This logic was added for MB-50333, because there exists a sanity
check in VBucket::abort() for replicas, which is done when the prepare
seqno of an abort is different from the one that is stored for the
prepare in the HT. This can happen when the prepare was deduped, but
such dedupe can happen only on disk checkpoints, as in-memory, prepares
and aborts go in different checkpoints.

Even if disk dedupe happened, the prepare seqno should be in the same
disk snapshot. This is what the assertion in abort processing checks.
To satisfy it, we ended up sending the start seqno of the checkpoint,
instead of that of the first mutation. We did this for both disk and
in-memory checkpoints, for consistency.

However, since for a new ActiveStream, the lastSnapEndSeqno is initially
0, we always end up taking the branch which pins the snapStartSeqno for
the very first SnapshotMarker. This is not correct in the case where the
consumer has reconnected, and we can immediately start streaming from
memory, and where we start streaming from memory from a checkpoint where
previous mutations have been expelled.

Example:

StreamRequest{snapshot=[5,5], seqno=[5, inf]}
                                    ^^^
                      next mutation to send is seqno:6
cursor
  v
[e:1, s:1, <expelled>, m:6, e:7]

In this case, where the next available mutation is the one we need,
and it is the first mutation in the checkpoint, we register the cursor
at the empty item. Due to the branch added in MB-50333, we would send
snapshot marker [1,6] and not backfill, as the item we need is
available.

We should consider the requested snapshot start_seqno, as when a stream
starts in the middle of a checkpoint and all items up to that point
have been deduped or expelled, the checkpoint start seqno can be before
the stream start_seqno, causing failures in the consumer. See
PassiveStream::reconnectStream where we set the expectation on the next
snapshot start.

While we could relax the expectation in PassiveStream instead, I believe
the correct thing to do is ensure that we don't change our behaviour
when expel or dedup has occurred, by picking a lower start seqno.

Change-Id: I9321ca5113e9a12eba0d2963cd97c273373daa29
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/203990
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>